executor: executor moved to it's own package
this is done in preperation for namespaced executors
